view override report fixes
